refactor the API
